+++ crates/rasteroid/src/lib.rs
@@ -85,3 +85,10 @@ impl InlineEncoder {
         Self::Ascii
     }
 }
+
+pub trait Frame {
+    fn width(&self) -> u16;
+    fn height(&self) -> u16;
+    fn timestamp(&self) -> f32;
+    fn data(&self) -> &[u8];
+}